Remove types/acp-types.ts bridge file and complete domain model migration
