+++ src/components/chat/MessageContentRenderer.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 const { useMemo } = React;
-import type { MessageContent, IAcpClient } from "../../types/acp-types";
+import type { MessageContent } from "../../domain/models/chat-message";
+import type { IAcpClient } from "../../adapters/acp.adapter";
 import type AgentClientPlugin from "../../main";
 import type { HandlePermissionUseCase } from "../../use-cases/handle-permission.use-case";
 import { MarkdownTextRenderer } from "./MarkdownTextRenderer";
